Fix up namespaces post-rename
