+++ Solutions/Corvus.SpecFlow.Extensions.Demo/AzureFunctionsTesting/DemoFunctionPerFeatureHooks.cs
@@ -5,9 +5,10 @@
 namespace Corvus.SpecFlow.Extensions.Demo.AzureFunctionsTesting
 {
     using System.Threading.Tasks;
-    using Corvus.AzureFunctions;
     using Corvus.SpecFlow.Extensions;
     using TechTalk.SpecFlow;
+    using Corvus.Testing.AzureFunctions;
+    using Corvus.Testing.AzureFunctions.Internal;
 
     [Binding]
     public static class DemoFunctionPerFeatureHooks
