+++ Solutions/Corvus.SpecFlow.Extensions.Demo/AzureFunctionsTesting/DemoFunctionPerScenarioHooks.cs
@@ -5,8 +5,8 @@
 namespace Corvus.SpecFlow.Extensions.Demo.AzureFunctionsTesting
 {
     using System.Threading.Tasks;
-    using Corvus.AzureFunctions;
     using Corvus.SpecFlow.Extensions;
+    using Corvus.Testing.AzureFunctions.Internal;
     using TechTalk.SpecFlow;
 
     [Binding]
