+++ Solutions/Corvus.SpecFlow.Extensions/Corvus/SpecFlow/Extensions/FunctionConfigurationExtensions.cs
@@ -5,7 +5,7 @@
 namespace Corvus.SpecFlow.Extensions
 {
     using System.Collections.Generic;
-    using Corvus.AzureFunctions;
+    using Corvus.Testing.AzureFunctions.Internal;
     using Microsoft.Extensions.Configuration;
     using TechTalk.SpecFlow;
