+++ Solutions/Corvus.Testing.AzureFunctions/Corvus/Testing/AzureFunctions/Internal/ProcessOutputHandler.cs
@@ -2,7 +2,7 @@
 // Copyright (c) Endjin Limited. All rights reserved.
 // </copyright>
 
-namespace Corvus.AzureFunctions.Internal
+namespace Corvus.Testing.AzureFunctions.Internal
 {
     using System;
     using System.Diagnostics;
